pyproject.toml
[project]
name = "ishop"
version = "0.1.0"
description = "iShop Api"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-decouple>=3.8",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
]

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
extend-select = ["I", "Q"]

[tool.ruff.format]
quote-style = "double"